requirements.txt
google-api-python-client>=2.156.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
yt-dlp>=2024.12.13
moviepy>=1.0.3
python-dotenv>=1.0.1
decorator>=5.1.1
imageio>=2.36.1
imageio-ffmpeg>=0.5.1
numpy>=2.2.1
proglog>=0.1.10
pillow>=10.4.0
tqdm>=4.67.1
requests>=2.32.3
urllib3>=2.3.0
certifi>=2024.12.14
psutil>=5.9.8
ffmpeg-python>=0.2.0